Moved files for creating filtering options into their own folder
